requirements.txt
openai
python-docx
pdfminer.six
python-dotenv
pydantic
google-generativeai
